Move health plugin body to new location
