+++ rpm/tarantool-metrics.spec
@@ -44,8 +44,6 @@ cp -rv cartridge %{br_luapkgdir}
      %{luapkgdir}/metrics/plugins/graphite/init.lua
 %dir %{luapkgdir}/metrics/plugins/json
      %{luapkgdir}/metrics/plugins/json/init.lua
-%dir %{luapkgdir}/metrics/plugins/health
-     %{luapkgdir}/metrics/plugins/health/health.lua
 %dir %{luapkgdir}/metrics/default_metrics
 %dir %{luapkgdir}/metrics/default_metrics/tarantool
      %{luapkgdir}/metrics/default_metrics/tarantool/fibers.lua
@@ -69,6 +67,7 @@ cp -rv cartridge %{br_luapkgdir}
 %dir %{luapkgdir}/metrics/tarantool
      %{luapkgdir}/metrics/tarantool/luajit.lua
 %dir %{luapkgdir}/cartridge
+     %{luapkgdir}/cartridge/health.lua
 %dir %{luapkgdir}/cartridge/roles
      %{luapkgdir}/cartridge/roles/metrics.lua
